Refactor WelcomeViewController environment initialization
